requirements.txt
fastapi
sqlalchemy
pymysql
pydantic
uvicorn
python-dotenv
fastapi-mail
jinja2